# Implement fast `rep_int()` for use in joins