@hadley 2020-02-24
I think I'd prefer this go in vctrs, even if it doesn't make it to dplyr in this release.