@romainfrancois 2020-02-25
I naively tried `vec_c(!!!map2(x, times, vec_recycle))` assuming the result of `vec_recycle()` would benefit altrep,  but no good on that example: 

``` r
library(dplyr, warn.conflicts = FALSE)
library(purrr)
library(vctrs)

rep_int <- dplyr:::rep_int

set.seed(123)

x <- 1:1e3 + 0L
times <- sample(1e3, replace = TRUE) + 0L

bench::mark(
  a = rep(x, times),
  b = rep.int(x, times),
  c = rep_int(x, times),
  d = vec_c(!!!map2(x, times, vec_recycle), .ptype = integer()),
  iterations = 1000
)
#> # A tibble: 4 x 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 a           694.1µs 930.69µs      885.     1.9MB     37.8
#> 2 b          809.15µs 836.63µs      963.    1.91MB     42.2
#> 3 c          188.84µs 274.41µs     2153.     1.9MB     94.4
#> 4 d            1.17ms   1.38ms      709.    3.87MB     93.9
```

<sup>Created on 2020-02-25 by the [reprex package](https://reprex.tidyverse.org) (v0.3.0.9000)</sup>

Perhaps if the loop was in C instead. `rlang` has `rep_along()`, maybe it could have `rep_each()` too or `vctrs::vec_rep_each()` ping @lionel-