@DavisVaughan 2020-02-25
@romainfrancois yea, the `map2()` is heavy there

``` r
set.seed(123)

x <- 1:1e3 + 0L
times <- sample(1e3, replace = TRUE) + 0L

library(purrr)

bench::mark(map2(x, times, function(x, y) {}))
#> # A tibble: 1 x 6
#>   expression                             min median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                         <bch:t> <bch:>     <dbl> <bch:byt>    <dbl>
#> 1 map2(x, times, function(x, y) { })   596µs  728µs     1376.    3.72MB     30.0
```

<sup>Created on 2020-02-25 by the [reprex package](https://reprex.tidyverse.org) (v0.3.0)</sup>

If we can work out the API for vctrs, it could live there. Maybe `vec_rep_each()` and `vec_rep_full()` as described here? https://principles.tidyverse.org/cs-rep.html#how-might-we-improve-the-situation