@hadley 2020-03-01
Closing, since I've linked to this implementation from the appropriate vctrs issue so we won't forget about it.